+++ Axis.Luna.Common/Results/ResultException.cs
@@ -1,12 +1,13 @@
 ﻿using System;
 using System.Diagnostics;
+using System.Runtime.ExceptionServices;
 
 namespace Axis.Luna.Common.Results
 {
     /// <summary>
     /// An exception created having a stack trace pointing to the site where this exception was created.
     /// Facilitates <see cref="IResult{TData}"/>s having stack traces that correspond to where the Error result was created.
-    /// This was born from the need to have stack traces in the result exception, expecially in cases where the result
+    /// This was born from the need to have stack traces in the result exception, especially in cases where the exception
     /// passed into the Result was not thrown prior.
     /// </summary>
     public class ResultException : Exception
@@ -16,7 +17,6 @@ namespace Axis.Luna.Common.Results
         {
             if (cause is null)
                 throw new ArgumentNullException(nameof(cause));
-
             this.OverwriteStackTrace(new StackTrace(1));
         }